server/requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
aiohttp-jwt==0.5.0
aiopg==1.0.0
alembic==1.4.2
async-timeout==3.0.1
attrs==19.3.0
brotlipy==0.7.0
cchardet==2.1.6
cffi==1.14.0
chardet==3.0.4
idna==2.9
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
multidict==4.7.5
psycopg2-binary==2.8.4
py-bcrypt==0.4
pycares==3.1.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
PyYAML==5.3.1
six==1.14.0
SQLAlchemy==1.3.15
yarl==1.4.2
